feat(#277): update api interface for paymentkeyhash module
