+++ src/GeniusYield/Types/PaymentKeyHash.hs
@@ -15,4 +15,40 @@ module GeniusYield.Types.PaymentKeyHash (
   paymentKeyHashToLedger,
 ) where
 
+import Cardano.Api qualified as Api
+import Cardano.Api.Ledger qualified as Ledger
+import GeniusYield.Imports
 import GeniusYield.Types.Hash
+import GeniusYield.Types.KeyRole
+import GeniusYield.Types.Ledger
+import GeniusYield.Types.PubKeyHash (
+  AsPubKeyHash (..),
+  pubKeyHashFromPlutus,
+  pubKeyHashToPlutus,
+ )
+import PlutusLedgerApi.V1.Crypto qualified as Plutus
+
+-- | @type GYPaymentKeyHash = GYKeyHash 'GYKeyRolePayment@
+type GYPaymentKeyHash = GYKeyHash 'GYKeyRolePayment
+
+paymentKeyHashFromPlutus :: Plutus.PubKeyHash -> Either PlutusToCardanoError GYPaymentKeyHash
+paymentKeyHashFromPlutus = fmap fromPubKeyHash . pubKeyHashFromPlutus
+
+paymentKeyHashToPlutus :: GYPaymentKeyHash -> Plutus.PubKeyHash
+paymentKeyHashToPlutus = toPubKeyHash >>> pubKeyHashToPlutus
+
+{-# DEPRECATED paymentKeyHashToApi "Use keyHashToApi." #-}
+paymentKeyHashToApi :: GYPaymentKeyHash -> Api.Hash Api.PaymentKey
+paymentKeyHashToApi = keyHashToApi
+
+{-# DEPRECATED paymentKeyHashFromApi "Use keyHashFromApi." #-}
+paymentKeyHashFromApi :: Api.Hash Api.PaymentKey -> GYPaymentKeyHash
+paymentKeyHashFromApi = keyHashFromApi
+
+{-# DEPRECATED paymentKeyHashToLedger "Use keyHashToLedger." #-}
+paymentKeyHashToLedger :: GYPaymentKeyHash -> Ledger.KeyHash Ledger.Payment Ledger.StandardCrypto
+paymentKeyHashToLedger = keyHashToLedger
+
+{-# DEPRECATED paymentKeyHashFromLedger "Use keyHashFromLedger." #-}
+paymentKeyHashFromLedger :: Ledger.KeyHash Ledger.Payment Ledger.StandardCrypto -> GYPaymentKeyHash
+paymentKeyHashFromLedger = keyHashFromLedger
